Move CommandlineOptions Hander Directory
