+++ TSQLLint.Tests/IntegrationTests/Configuration/ExistingConfigTests.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using NUnit.Framework;
-using TSQLLint.Console.ConfigHandler;
+using TSQLLint.Console.CommandLineOptions;
 using TSQLLint.Lib.Rules.RuleViolations;
 
 namespace TSQLLint.Tests.IntegrationTests
